# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * DevHub personal workspace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Note {
  id: number;
  title: string;
  content: string;
  tags: string[];
  pinned: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface NoteInput {
  /** @minLength 1 */
  title: string;
  content?: string;
  tags?: string[];
  pinned?: boolean;
}

export interface NoteUpdate {
  /** @minLength 1 */
  title?: string;
  content?: string;
  tags?: string[];
  pinned?: boolean;
}

export interface Reminder {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** ISO-8601 datetime */
  dueAt: string;
  done: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface ReminderInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  dueAt: string;
  done?: boolean;
}

export interface ReminderUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  dueAt?: string;
  done?: boolean;
}

export interface Meeting {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** ISO-8601 datetime */
  startAt: string;
  /**
     * ISO-8601 datetime
     * @nullable
     */
  endAt?: string | null;
  /** @nullable */
  meetLink?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface MeetingInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  startAt: string;
  endAt?: string;
  meetLink?: string;
}

export interface MeetingUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  startAt?: string;
  endAt?: string;
  meetLink?: string;
}

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  todo: 'todo',
  in_progress: 'in_progress',
  done: 'done',
} as const;

export type TaskPriority = typeof TaskPriority[keyof typeof TaskPriority];


export const TaskPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  /** @nullable */
  dueAt?: string | null;
  tags?: string[];
  createdAt: string;
  updatedAt: string;
}

export type TaskInputStatus = typeof TaskInputStatus[keyof typeof TaskInputStatus];


export const TaskInputStatus = {
  todo: 'todo',
  in_progress: 'in_progress',
  done: 'done',
} as const;

export type TaskInputPriority = typeof TaskInputPriority[keyof typeof TaskInputPriority];


export const TaskInputPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  status?: TaskInputStatus;
  priority?: TaskInputPriority;
  dueAt?: string;
  tags?: string[];
}

export type TaskUpdateStatus = typeof TaskUpdateStatus[keyof typeof TaskUpdateStatus];


export const TaskUpdateStatus = {
  todo: 'todo',
  in_progress: 'in_progress',
  done: 'done',
} as const;

export type TaskUpdatePriority = typeof TaskUpdatePriority[keyof typeof TaskUpdatePriority];


export const TaskUpdatePriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  status?: TaskUpdateStatus;
  priority?: TaskUpdatePriority;
  dueAt?: string;
  tags?: string[];
}

export interface Prompt {
  id: number;
  title: string;
  content: string;
  /** e.g. gpt-4, claude-3, gemini-pro */
  model: string;
  tags: string[];
  createdAt: string;
  updatedAt: string;
}

export interface PromptInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  content: string;
  model?: string;
  tags?: string[];
}

export interface PromptUpdate {
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  content?: string;
  model?: string;
  tags?: string[];
}

export interface Snippet {
  id: number;
  title: string;
  code: string;
  /** e.g. typescript, python, bash */
  language: string;
  /** @nullable */
  description?: string | null;
  tags: string[];
  createdAt: string;
  updatedAt: string;
}

export interface SnippetInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  language: string;
  description?: string;
  tags?: string[];
}

export interface SnippetUpdate {
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  code?: string;
  /** @minLength 1 */
  language?: string;
  description?: string;
  tags?: string[];
}

export interface Bookmark {
  id: number;
  title: string;
  url: string;
  /** @nullable */
  description?: string | null;
  tags: string[];
  createdAt: string;
  updatedAt: string;
}

export interface BookmarkInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  url: string;
  description?: string;
  tags?: string[];
}

export interface BookmarkUpdate {
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  url?: string;
  description?: string;
  tags?: string[];
}

export type DashboardSummaryCounts = {
  notes: number;
  reminders: number;
  meetings: number;
  tasks: number;
  prompts: number;
  snippets: number;
  bookmarks: number;
};

export interface DashboardSummary {
  counts: DashboardSummaryCounts;
  upcomingReminders: Reminder[];
  todayMeetings: Meeting[];
  recentNotes: Note[];
  pendingTasks: Task[];
}