# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DevHub personal workspace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List all notes
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    get:
      operationId: getNote
      tags: [notes]
      summary: Get a note by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Reminders ────────────────────────────────────────────────────────────
...
  /meetings:
    get:
      operationId: listMeetings
      tags: [meetings]
      summary: List all meetings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meeting"
    post:
      operationId: createMeeting
      tags: [meetings]
      summary: Create a meeting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
  /meetings/{id}:
    get:
      operationId: getMeeting
      tags: [meetings]
      summary: Get a meeting by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
        "404":
          description: Not found
    patch:
      operationId: updateMeeting
      tags: [meetings]
      summary: Update a meeting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetingUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
        "404":
          description: Not found
    delete:
      operationId: deleteMeeting
      tags: [meetings]
      summary: Delete a meeting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Tasks ─────────────────────────────────────────────────────────────────
...
  /prompts:
    get:
      operationId: listPrompts
      tags: [prompts]
      summary: List all AI prompts
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prompt"
    post:
      operationId: createPrompt
      tags: [prompts]
      summary: Create an AI prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromptInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
  /prompts/{id}:
    get:
      operationId: getPrompt
      tags: [prompts]
      summary: Get a prompt by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
        "404":
          description: Not found
    patch:
      operationId: updatePrompt
      tags: [prompts]
      summary: Update a prompt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromptUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
        "404":
          description: Not found
    delete:
      operationId: deletePrompt
      tags: [prompts]
      summary: Delete a prompt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Code Snippets ─────────────────────────────────────────────────────────
...
  /bookmarks:
    get:
      operationId: listBookmarks
      tags: [bookmarks]
      summary: List all bookmarks
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bookmark"
    post:
      operationId: createBookmark
      tags: [bookmarks]
      summary: Create a bookmark
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookmarkInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bookmark"
  /bookmarks/{id}:
    get:
      operationId: getBookmark
      tags: [bookmarks]
      summary: Get a bookmark by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bookmark"
        "404":
          description: Not found
    patch:
      operationId: updateBookmark
      tags: [bookmarks]
      summary: Update a bookmark
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookmarkUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bookmark"
        "404":
          description: Not found
    delete:
      operationId: deleteBookmark
      tags: [bookmarks]
      summary: Delete a bookmark
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Dashboard ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Note:
      type: object
      required: [id, title, content, tags, pinned, createdAt, updatedAt]
      properties:
        id:        { type: integer }
        title:     { type: string }
        content:   { type: string }
        tags:      { type: array, items: { type: string } }
        pinned:    { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
    NoteInput:
      type: object
      required: [title]
      properties:
        title:   { type: string, minLength: 1 }
        content: { type: string }
        tags:    { type: array, items: { type: string } }
        pinned:  { type: boolean }
    NoteUpdate:
      type: object
      properties:
        title:   { type: string, minLength: 1 }
        content: { type: string }
        tags:    { type: array, items: { type: string } }
        pinned:  { type: boolean }

    # ── Reminder ──────────────────────────────────────────────────────────
...
    Meeting:
      type: object
      required: [id, title, startAt, createdAt, updatedAt]
      properties:
        id:          { type: integer }
        title:       { type: string }
        description: { type: ["string", "null"] }
        startAt:     { type: string, description: "ISO-8601 datetime" }
        endAt:       { type: ["string", "null"], description: "ISO-8601 datetime" }
        meetLink:    { type: ["string", "null"] }
        createdAt:   { type: string }
        updatedAt:   { type: string }
    MeetingInput:
      type: object
      required: [title, startAt]
      properties:
        title:       { type: string, minLength: 1 }
        description: { type: string }
        startAt:     { type: string }
        endAt:       { type: string }
        meetLink:    { type: string }
    MeetingUpdate:
      type: object
      properties:
        title:       { type: string, minLength: 1 }
        description: { type: string }
        startAt:     { type: string }
        endAt:       { type: string }
        meetLink:    { type: string }

    # ── Task ──────────────────────────────────────────────────────────────
...
    Prompt:
      type: object
      required: [id, title, content, model, tags, createdAt, updatedAt]
      properties:
        id:        { type: integer }
        title:     { type: string }
        content:   { type: string }
        model:     { type: string, description: "e.g. gpt-4, claude-3, gemini-pro" }
        tags:      { type: array, items: { type: string } }
        createdAt: { type: string }
        updatedAt: { type: string }
    PromptInput:
      type: object
      required: [title, content]
      properties:
        title:   { type: string, minLength: 1 }
        content: { type: string, minLength: 1 }
        model:   { type: string }
        tags:    { type: array, items: { type: string } }
    PromptUpdate:
      type: object
      properties:
        title:   { type: string, minLength: 1 }
        content: { type: string, minLength: 1 }
        model:   { type: string }
        tags:    { type: array, items: { type: string } }

    # ── Snippet ───────────────────────────────────────────────────────────
...
    Bookmark:
      type: object
      required: [id, title, url, tags, createdAt, updatedAt]
      properties:
        id:          { type: integer }
        title:       { type: string }
        url:         { type: string }
        description: { type: ["string", "null"] }
        tags:        { type: array, items: { type: string } }
        createdAt:   { type: string }
        updatedAt:   { type: string }
    BookmarkInput:
      type: object
      required: [title, url]
      properties:
        title:       { type: string, minLength: 1 }
        url:         { type: string, minLength: 1 }
        description: { type: string }
        tags:        { type: array, items: { type: string } }
    BookmarkUpdate:
      type: object
      properties:
        title:       { type: string, minLength: 1 }
        url:         { type: string, minLength: 1 }
        description: { type: string }
        tags:        { type: array, items: { type: string } }

    # ── Dashboard ─────────────────────────────────────────────────────────
...