# lib/api-zod/src/generated/types/bookmark.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * DevHub personal workspace API
 * OpenAPI spec version: 0.1.0
 */

export interface Bookmark {
  id: number;
  title: string;
  url: string;
  /** @nullable */
  description?: string | null;
  tags: string[];
  createdAt: string;
  updatedAt: string;
}