# lib/api-zod/src/generated/types/promptInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * DevHub personal workspace API
 * OpenAPI spec version: 0.1.0
 */

export interface PromptInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  content: string;
  model?: string;
  tags?: string[];
}